Move orthogonalize to utils.py
