+++ mlbench_core/optim/pytorch/optim.py
@@ -4,6 +4,7 @@ from mlbench_core.utils.pytorch.distributed import (
     pack_tensors,
     unpack_tensors,
 )
+from mlbench_core.optim.pytorch.utils import orthogonalize
 
 import numpy as np
 import torch
@@ -553,7 +554,6 @@ class PowerSGD(Optimizer):
     def set_random(self, vector):
         torch.manual_seed(self.rng.randint(1_000_000_000))
         vector.data[:] = torch.randn(*vector.shape, device=self.device)
-        # orthogonalize(vector)
 
     def step(self, closure=None):
         """Performs a single optimization step.
@@ -665,7 +665,6 @@ class PowerSGD(Optimizer):
             n, m = matrix.shape
 
             if self.reuse_query and not memory_is_uninitialized:
-                # orthogonalize(q)
                 pass
             else:
                 # Sample a query vector q
@@ -702,20 +701,6 @@ class PowerSGD(Optimizer):
         rank1_unpacked = unpack_tensors(rank1_packed, rank1_indices, rank1_sizes)
         for i, (_, out, _) in enumerate(rank1_tensors):
             out[:] = rank1_unpacked[i]
-
-
-@torch.jit.script
-def orthogonalize(matrix):
-    n, m = matrix.shape
-    for i in range(m):
-        # Normalize the i'th column
-        col = matrix[:, i : i + 1]
-        col /= torch.sqrt(torch.sum(col ** 2))
-        # Project it on the rest and remove it
-        if i + 1 < m:
-            rest = matrix[:, i + 1 :]
-            # rest -= torch.matmul(col.t(), rest) * col
-            rest -= torch.sum(col * rest, dim=0) * col
 
 
 optimizers = {
