+++ mlbench_core/optim/pytorch/utils.py
@@ -1,0 +1,14 @@
+import torch
+
+@torch.jit.script
+def orthogonalize(matrix):
+    n, m = matrix.shape
+    for i in range(m):
+        # Normalize the i'th column
+        col = matrix[:, i : i + 1]
+        col /= torch.sqrt(torch.sum(col ** 2))
+        # Project it on the rest and remove it
+        if i + 1 < m:
+            rest = matrix[:, i + 1 :]
+            # rest -= torch.matmul(col.t(), rest) * col
+            rest -= torch.sum(col * rest, dim=0) * col
